requirements.txt
torch==2.1.2+cu118
numpy==1.24.3
tqdm==4.66.1
einops==0.7.0
torchvision==0.16.2+cu118
omegaconf==2.3.0
timm==0.9.12
opencv-python==4.8.1.78
transformers==4.36.2
scipy==1.11.4
albumentations==1.3.1
Pillow==9.3.0
pip==21.3.1
toml==0.10.2
wheel==0.37.1
tornado==6.4
setuptools==60.2.0
future==0.18.3
requests==2.28.2
lemon==1.3.1
matplotlib==3.8.2
scikit-image==0.22.0
regex==2023.12.25
tifffile==2023.12.9
addict==2.4.0
yapf==0.40.2
packaging==23.2
PyYAML==6.0.1
tensorboard==2.15.1
prettytable==3.9.0
pandas==2.0.1
latextable==1.0.1
texttable==1.7.0
gradio==4.12.0